requirements.txt
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
rank-bm25>=0.2.2
streamlit>=1.35.0
google-genai>=1.0.0
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
tqdm>=4.66.0
